clean up project structure; fix: breadcrumbs
