+++ client/src/App.tsx
@@ -1,13 +1,13 @@
 import React from "react";
-import Header from "./Header";
-import HomePage from "../pages/HomePage";
-import TestScheduler from "./TestScheduler";
-import DegNav from "./DegNav";
-import AllCourses from "./AllCourses";
+import Header from "./components/Header";
+import HomePage from "./pages/HomePage";
+import TestScheduler from "./pages/TestScheduler";
+import DegNav from "./pages/DegNav";
+import AllCourses from "./pages/AllCourses";
 import { BrowserRouter, Route } from "react-router-dom";
 import { Provider } from "react-redux";
 import thunk from "redux-thunk";
-import reducers from "../reducers";
+import reducers from "./reducers";
 import { createStore, applyMiddleware } from "redux";
 import { StylesProvider, ThemeProvider, createMuiTheme } from "@material-ui/core/styles";
 import CssBaseline from "@material-ui/core/CssBaseline";
