+++ client/src/pages/AllCourses.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const AllCourses = () => {
+  return (
+    <div>
+      <h3>Welcome to the courses page!</h3>
+    </div>
+  );
+};
+
+export default AllCourses;
