+++ client/src/pages/DegNav.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const DegNavPage = () => {
+  return (
+    <div>
+      <h3>This is the Degree Navigator Page!</h3>
+    </div>
+  );
+};
+
+export default DegNavPage;
